Properly pass bindgen cflags to bindgen.
This gets me a cross build without anything else.

It's probable that we can simplify a bunch of the stuff in the build.rs file
with this patch, but I haven't tried to do that yet.

Whitespace handling is a bit broken (I hope nobody tries to build this crate in
a directory with spaces) but all the other uses of split_whitespace are equally
broken in build.rs.