Also respect readelf and objcopy cc-rs variants
While compiling servo for OpenHarmony on a MacOS host,
I noticed that readelf and objcopy apparently aren't chosen
from the same directory as the clang c compiler, but needed
to be in PATH for the configure script to pick them up.
To allow users to set READELF and OBJCOPY explicitly, we
should also read the relevant cc-rs environment variables.

An alternative would have been to inspect the configure scripts
and find out why they don't look next to clang, but I'm not
familiar enough with the configure script.