Arc should not be Traceable (#561)
Traceable works under the assumption that only one thread will be
looking at values. The entire purpose of `Arc` is to share things
between threads. This ~never works (unless the Arc could be replaced by
a Rc).

Nothing in this crate or servo uses this impl. Let's just delete this.

Signed-off-by: Greg Morenz <greg-morenz@droid.cafe>