ohos: Use custom domain when logging with hilog
This modifies the existing patch to redirect log messages to
hilog to use a non-zero domain, which allows us to filter
logs via the domain on ohos devices.
We set 0xE0C3 for Rust code in servo, using a different number
allows us to seperate spidermonkey from servo logs if wanted.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>

# Conflicts:
#	mozjs-sys/etc/patches/0036-Redirect_errors_to_hilog_on_OpenHarmony.patch
#	mozjs-sys/mozjs/memory/mozalloc/mozalloc_abort.cpp
#	mozjs-sys/mozjs/mozglue/misc/Debug.cpp
#	mozjs-sys/mozjs/nsprpub/pr/src/io/prlog.c