ohos: Use custom domain when logging with hilog
This modifies the existing patch to redirect log messages to
hilog to use a non-zero domain, which allows us to filter
logs via the domain on ohos devices.
We set 0xE0C3 for Rust code in servo, using a different number
allows us to seperate spidermonkey from servo logs if wanted.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>