Improvements to the speed of latin1_to_string. (#587)
On the benchmark it improves the encoding from around 9 microsends to 6
microseconds.

This PR also includes the benchmark and setup as criterion benchmark.

Also test the higher parts of latin1 which are not the same.

At the moment we have to use encoding_rs to do this encoding leading us
to a bit of a cyclical testing. But tests should prepare for future
refactoring, so I think this is fine.

Signed-off-by: Narfinger <Narfinger@users.noreply.github.com>