Use Rust 1.85 toolchain (#611)
* Use Rust 1.85 toolchain

We should ensure mozjs doesn't break servos MSRV accidentally.
The easiest way is to just consistently use servos MSRV as the
compiler when developing mozjs.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>

* Add scheduled runs with beta and stable

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>

* Explicitly use selected toolchain.

To ensure we always use the toolchain given by
input, we use the output of the rust-toolchain action,
which is guaranteed to be usable in the `cargo +<toolchain>`
form.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>

* Fix release-check rust toolchain input.

Consider `on: release` and add a rust_version input.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>

---------

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>