Prepare mozjs-sys for publishing (#617)
* Update mozjs_sys publishing excludes

Most of the previous folders don't exist anymore.
`testing` is 60MiB uncompressed, so we don't want
to ship that.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>

* Prevent python bytecode cache

Cargo will complain when publishing if the source directory
is polluted during build.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>

* Remove old-configure from makefile.cargo

The file doesn't exist anymore, so we shouldn't touch it.
This prevents cargo publish from complaining about failed
package verification.

Cargo will complain when publishing if the source directory
is polluted during build.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>

* Update authors for mozjs-sys

Add the servo project developers to the authors.
This should make it more clear that this is not an official mozilla crate.
We don't remove Mozilla from the list, since they are the primary
authors of spidermonkey, aka. the main content of this crate.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>

* filter repo instead of excluding on publish

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>

* bump version

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>

---------

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>