Support build configurations without Intl API (#623)
* Support build configurations without Intl API

The Intl API requires bundling a very large amount of data, and isn't needed in absolutely all environments.

This commit introduces an `intl` feature, which is enabled by default on `mozjs`, and builds SpiderMonkey without Intl support if that feature is disabled.

On MacOS, a mozjs embedding I'm working on shrinks from 32MB to 12MB, and a wasm32 build from also 32MB to 8.5MB (all stripped).

Signed-off-by: Till Schneidereit <till@tillschneidereit.net>

* Always build from source if intl feature is disabled

Signed-off-by: Till Schneidereit <till@tillschneidereit.net>

* Bump `mozjs-sys` version

Signed-off-by: Till Schneidereit <till@tillschneidereit.net>

---------

Signed-off-by: Till Schneidereit <till@tillschneidereit.net>