Remove MutableHandle::new (#689)
This was unused except from from_marked_location. It was unsound when
called from from_marked_location because it converted the ptr to an &mut
reference invalidating the reference the GC holds. This was an oversight
in #572.

There is some discussion in https://github.com/servo/mozjs/pull/688#issuecomment-3689737344 
that from_marked_location might also be being called with null pointers.
I don't think that's happening based on a quick test with putting an
assert in and running servo's tests, and if it is we should consider
changing the API so that it doesn't happen in the future. In the
meantime this makes that sound as a side benefit (so long as the
MutableHandle isn't then used to access the value).

Nothing in the public API changes.

Signed-off-by: Greg Morenz <greg-morenz@droid.cafe>